fix compiler warnings

+++ caldgemm.cpp
@@ -529,13 +529,13 @@ int caldgemm::cpuScheduler()
 					if (cParam.dynamic_size && (DGEMM_favor_m ? gpu_m : gpu_n) % Config->Height)
 					{
 						const size_t adjustment = Config->Height - (DGEMM_favor_m ? gpu_m : gpu_n) % Config->Height;
-						fprintf(STD_OUT, "Adjusting second phase run size for small tiles: %lld - %lld = %lld\n", cParam.dynamic_size, adjustment, cParam.dynamic_size - adjustment);
+						fprintf(STD_OUT, "Adjusting second phase run size for small tiles: %lld - %lld = %lld\n", (long long int) cParam.dynamic_size, (long long int) adjustment, (long long int) cParam.dynamic_size - adjustment);
 						cParam.dynamic_size -= adjustment;
 					}
 					if (cParam.dynamic_run && (DGEMM_favor_m ? gpu_n : gpu_m) % Config->Height)
 					{
 						const size_t adjustment = Config->Height - (DGEMM_favor_m ? gpu_n : gpu_m) % Config->Height;
-						fprintf(STD_OUT, "Adjusting second phase run row count for small tiles: %lld - %lld = %lld\n", cParam.dynamic_run, adjustment, cParam.dynamic_run - adjustment);
+						fprintf(STD_OUT, "Adjusting second phase run row count for small tiles: %lld - %lld = %lld\n", (long long int) cParam.dynamic_run, (long long int) adjustment, (long long int) cParam.dynamic_run - adjustment);
 						cParam.dynamic_run -= adjustment;
 					}
